[electrum] handle 0x5515 status when a Ledger device is locked in screensaver mode
Summary:
When the BCH app is open for a while and the device goes to screensaver mode, and Electrum ABC is restarted with the Ledger wallet, it would print an unhelpful 'invalid status 5515' error. To recover, the user would have to unlock the device and close/reopen the wallet.
Handle this status with a more useful error dialog.

Test Plan: Test the above scenario

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16360