[ecash-lib] Export slpAmount function
Summary:
In implementing NFT trading in Cashtab, it is useful for this function to be available so Cashtab can validate the script of an advertised NFT.

I can see how this type of validation could be useful for other app dev use cases. There is not really a cost in exporting the function, as webpack will only include it in apps that use it.

Test Plan:
CI to confirm `ecash-lib` not affected.

I have not uploaded my latest working copy of D16328 yet, but locally, this change allows me to use the function in Cashtab, whereas without this change the function is not available.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16379