validation: Check if mempool exists before asserting in ActivateSnapshot
Summary:
This is mostly for consistency with how the mempool is accessed in other parts of the codebase. For now we don't expect the active chainstate to not have a mempool.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/30388 | core#30388]]
with a bugfix from https://github.com/bitcoin/bitcoin/pull/33728/changes/2bc32656498517fe58bd41dcbd0afd306d51d4b0

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20622