[Cashtab] Standardize implementation of token dust change outputs
Summary:
Have a working implementation of NFT trading in Cashtab now with D16328

Requires a number of new and more complicated transactions. To properly support these transactions, some improvements are required in how transaction inputs and outputs are prepared.

This change improves code readability and ensures consistency with the handling of token dust change outputs, i.e. a targetOutput created by Cashtab as a token UTXO intended to be received by the same wallet that broadcasts the transaction.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16425