Fix release notes for v0.29.8 (archive) and v0.29.9 (current)
Summary:
In D16444, a release note has been added but at the same time version number was decreased by error. This caused the bot to figure out the release notes needed to be archived and bumped because the CMakeLists.txt was on the right version, and it happened in rABC7a54d5acb5b9cca1851591a05bc1b30c7841dcad.

This diff fixes this issue by restauring the archive and the current release notes, this time with the correct version number.

Test Plan: Read the docs, check the versions, look at the git history to figure out what happened.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D16487