refactor: Implement strict uint256::FromHex()
Summary:
This is a safe replacement of the previous SetHex, which now returns an
optional to indicate success or failure.

The code is similar to the ParseHashStr helper, which will be removed in
a later commit.

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/30482 | core#30482]]
https://github.com/bitcoin/bitcoin/pull/30482/changes/fad2991ba073de0bd1f12e42bf0fbaca4a265508
Depends on D20680

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20681