[Cashtab] Show option to claim token reward from Home page if wallet is new
Summary:
A newly-created wallet is always eligible for token rewards.

Add an option to the first screen a user sees when they create a wallet to "Claim Airdrop."

Considered (and implemented in testing) the option of just having the wallet auto-claim an airdrop on creation. However this could be a bit confusing to the user. The "Backup wallet" msg flashes briefly. The airdrop received msg is nice, but it is unclear why or what happened.

Having the user click a button and shortly after receive a tx is imo more satisfying and less ambiguous.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Subscribers: emack

Differential Revision: https://reviews.bitcoinabc.org/D16509