test: unittest chainstatemanager_args
Summary:
I've run into some challenges while trying to adapt uint256 backports to our codebase. One of the issues was not to break support for "0x...." hashes or shorter than 64 char strings  when parsing `-assumevalid`
And then I found this commit that tests exactly that

This is a partial backport of [[https://github.com/bitcoin/bitcoin/pull/30569 | core#30569]]
https://github.com/bitcoin/bitcoin/pull/30569/changes/85b7cbfcbe3f94770bdf73dedd8bda0193a44627

Note that there is a one line difference with the source material, because we are missing the backport that turns the return type of ApplyArgsManOptions from an `optional<bilingual_string>` to a `util::Result<void>`. That turns the logic around when testing for success for this function.

Test Plan: `ninja check`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D20673