[explorer] Render cashaddr for all inputs and outputs on transaction page
Summary:
The block explorer has logic which changes the address format of the outputScript depending on whether or not the input or output sent or received tokens in the given tx.

While it is arguably nice to see that some outputs are token outputs, imo this is already shown enough by showing the token icon and the amount. Forcing the address to render as an `etoken:` address is confusing.

Address conversion is still possible with the explorer on the address page.

Test Plan: `cargo run`, no errors, check a token tx, note all inputs and outputs have `ecash:` prefixed addresses on the tx page

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16610