[explorer] Render token icon next to token inputs, outputs, and in token summary table
Summary:
Render the token icon next to token inputs and outputs, including the mint baton. Also render in the token summary table and the `eTokens` table on the address page.

Note: a blockie is generated server-side for tokens that do not have an icon.

Test Plan:
`cargo run`, check out the genesis tx for some tokens with and without icons.

{F8259896}

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16611