[mock-chronik-client] Bump version
Summary:
The last diff on mock chronik client was landed in a stack, and I missed bumping the version. See D16382

Bump the version so we get back on track.

Test Plan:
Proofread, confirm this was indeed missed.

Check README to note D16382 is mentioned as the diff for version 1.10.0. We should leave this link as that is the diff that added the new feature.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16616