[chronik client + ecash-herald + token-server + cashtab + ecash-lib + ecash-agora] Deprecate NNG support from chronik-client
Summary:
chronik was first developed on a bitcoin abc fork. It was later improved to work directly with the abc node in the monorepo.

This improved version was planned to be the only / default version.

Get rid of nng support. Rename all "InNode" types to lose this special marker. Rename `ChronikClientNode` to `ChronikClient`

Bump chronik-client to 1.0.0

Update ecash-herald, token-server, and Cashtab to handle this new type.

Note: alias-server and apps/examples are still legacy chronik nng apps. alias-server will be migrated to in-node if it is to be used for aliases. app/examples needs some thought. Potentially should be deprecated as the integration tests in chronik-client are more exhaustive, accurate, and easier to maintain.

I considered a phased approach where, say, we do not change all the names. I think though it is best to just drop this stuff, since there is no plan to maintain nng support going forward, and full migration to in-node was always the plan.

Test Plan:
CI

To test existing implementations can still import and use `ChronikClientNode`

- `cd modules/chronik-client` and `npm run build`
- `cd cashtab` and `npm start`, send a tx, no API error
- In `cashtab/src/index.js`, replace `ChronikClient` with `ChronikClientNode`
- `npm start`, send a tx, no API error

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: emack, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D16627