Revert "[avalanche] Sort txs to poll by modified fee rate"
Summary:
This reverts commit 223c43df3eb5ac6fabed516e7ef34adcbf5f87af.

The sorting algorithm in itself works but is introducing a design issue:
it needs the mempool to be locked to ensure deterministic ordering. But
the lock is only help during a single comparison and not during the whole
sorting which can lead to issues. Holding the lock at callsite is not
really doable because of the std::variant, or would be cause a huge
performance hit.

For now let's revert this and keep it simple by always sorting by txid.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, roqqit

Reviewed By: roqqit

Differential Revision: https://reviews.bitcoinabc.org/D16637